Extract useEditField, useGetEvent and usePublishEvent hooks
