+++ src/pages/manage/movies/useGetEvent.ts
@@ -1,0 +1,10 @@
+import { useGetEventByIdQuery } from '@/hooks/api/events';
+
+const useGetEvent = ({ id, onSuccess }) => {
+  const getEventByIdQuery = useGetEventByIdQuery({ id }, { onSuccess });
+
+  // @ts-expect-error
+  return getEventByIdQuery?.data;
+};
+
+export { useGetEvent };
